Share esbuild module-resolution policy with the gallery bundle
`npm run gallery:build` failed on master: the gallery config restated
production's module-resolution policy but incompletely. Its `node:*`
external matched only prefixed builtins, so the bare forms that
@anthropic-ai/claude-agent-sdk imports (`child_process`, `crypto`, `fs`,
…) went unresolved, and it configured neither an SVG loader for the
agent-mode backend logos nor the `module` shim.

Give both bundles one shared policy instead of a second copy to drift:
`nodeBuiltinExternals` moves next to the `module` shim it complements,
and the SVGR loader becomes a sibling of wasmPlugin/nodeModuleShim.
The production bundle is byte-identical before and after.
